+++ UVtools.Core/Operations/OperationRepairLayers.cs
@@ -14,6 +14,13 @@ namespace UVtools.Core.Operations
 {
     public class OperationRepairLayers : Operation
     {
+        private bool _repairIslands = true;
+        private bool _repairResinTraps = true;
+        private bool _removeEmptyLayers = true;
+        private byte _removeIslandsBelowEqualPixelCount = 5;
+        private ushort _removeIslandsRecursiveIterations = 4;
+        private uint _gapClosingIterations = 1;
+        private uint _noiseRemovalIterations = 0;
         public override bool CanROI { get; set; } = false;
         public override string Title => "Repair layers and issues";
         public override string Description => null;
@@ -37,12 +44,49 @@ namespace UVtools.Core.Operations
             return new StringTag(sb.ToString());
         }
 
-        public bool RepairIslands { get; set; } = true;
-        public bool RepairResinTraps { get; set; } = true;
-        public bool RemoveEmptyLayers { get; set; } = true;
-        public byte RemoveIslandsBelowEqualPixelCount { get; set; } = 5;
-        public uint GapClosingIterations { get; set; } = 1;
-        public uint NoiseRemovalIterations { get; set; } = 0;
+        public bool RepairIslands
+        {
+            get => _repairIslands;
+            set => RaiseAndSetIfChanged(ref _repairIslands, value);
+        }
+
+        public bool RepairResinTraps
+        {
+            get => _repairResinTraps;
+            set => RaiseAndSetIfChanged(ref _repairResinTraps, value);
+        }
+
+        public bool RemoveEmptyLayers
+        {
+            get => _removeEmptyLayers;
+            set => RaiseAndSetIfChanged(ref _removeEmptyLayers, value);
+        }
+
+        public byte RemoveIslandsBelowEqualPixelCount
+        {
+            get => _removeIslandsBelowEqualPixelCount;
+            set => RaiseAndSetIfChanged(ref _removeIslandsBelowEqualPixelCount, value);
+        }
+
+        public ushort RemoveIslandsRecursiveIterations
+        {
+            get => _removeIslandsRecursiveIterations;
+            set => RaiseAndSetIfChanged(ref _removeIslandsRecursiveIterations, value);
+        }
+
+        public uint GapClosingIterations
+        {
+            get => _gapClosingIterations;
+            set => RaiseAndSetIfChanged(ref _gapClosingIterations, value);
+        }
+
+        public uint NoiseRemovalIterations
+        {
+            get => _noiseRemovalIterations;
+            set => RaiseAndSetIfChanged(ref _noiseRemovalIterations, value);
+        }
+
+        public IslandDetectionConfiguration IslandDetectionConfig { get; set; }
 
         public List<LayerIssue> Issues { get; set; }